move expectError into utils
