+++ src/app/components/SkipToContentWrapper.spec.tsx
@@ -2,6 +2,7 @@ import { HTMLElement } from '@openstax/types/lib.dom';
 import React, { ReactElement } from 'react';
 import ReactDOM from 'react-dom';
 import ReactTestUtils from 'react-dom/test-utils';
+import { expectError } from '../../test/reactutils';
 import MainContent from './MainContent';
 import SkipToContentWrapper from './SkipToContentWrapper';
 
@@ -16,23 +17,6 @@ function renderToDom<P>(component: ReactElement<P>) {
         throw new Error(`BUG: Could not find DOM node`);
     }
     return {component: c, node};
-}
-
-// JSDom logs to console.error when an Error is thrown.
-// Disable the console just in this instance, and re-enable after.
-// See https://github.com/facebook/jest/pull/5267#issuecomment-356605468
-function expectError(message: string, fn: () => void) {
-    const consoleError = jest.spyOn(console, 'error');
-    consoleError.mockImplementation(() => { /* suppress jsdom logging to console */ });
-
-    try {
-        fn();
-        expect('Code should not have succeeded').toBeFalsy();
-    } catch (err) {
-        expect(err.message).toBe(message);
-    } finally {
-        consoleError.mockRestore();
-    }
 }
 
 describe('SkipToContentWrapper', () => {
